Issue #62: improved view command to allow viewing of remote files.
Viewing of remote files where the url is provided as virtual url (grid://) is very slow because there's a lot of stuff involved for the backend to actually figure out the real url.
Also, disabled viewer type functionality Yuriy wrote for now, didn't work for me. We can add that later on again.